decoration panel logistics and UI almost complete
The selected decoration layer is stored and the UI is written to show which layer is selected with appropriate animations smooth and seamless. Tree View to view all editable properties is visible in a tree view with expanding animations, whichever property needs to be editted can be pinned to display its UI for editing, these pins are stores in database as well as a JSON object corresponds to every editable value in the itemDecoration class. The expansion of tree is not stored in the database but is stored locally for the as long the layout designer widget is not popped.