Hive now stores and retrives complex web of interrelated Decorations and super decorations
Yes Ideally I should've implemented the Hive Storage way earlier but I wasn't sure how I was going to approach the pinning and UI so there were still a few attributes left to add in the HiveBox classes. All of which is done now, the Decorations Hivebox stores SuperDecorations with its children, ItemDecoration on the other hand im not sure could be directly stored in the box but that is a bridge we'll cross soon. Next in line is the actual editor UI for tweaking and saving the parameters.