commit before storing decorations globally
The Ui for decoration Editor is almost done. But to search through decorations to find it and when dealing with times where rows or fields are copy pasted the decorations need to stay linked but instead only changes to one of them is stored in the database. while the app runs its no worries but database is crucial