the table decoration can be switched from row column table and tablebg.
updating the row decoration will update all cells in that row, respectively for column. same for table and table bg. The priority is cell decoration. If the cell has a decoration that is different from the table decoration and row or column decoration then the cell wont switch to using the row or column decoration.