itemlist, totalPayable added, bills and layouts are now separate and accesible.
bills can be made based on a layout. Bills /Invoices can be revised into a revised bill separate from the original. Changing the label of a text to match that of a required field will automatically format check the existing text if it matches the type requirement of the required text and handle it accordingly. The placeholder text is also updated on label change for each SheetTextType.

Itemlists are tables with headers that are mandatory requirements for the specific type of bill or invoice. I.e. Item Description, Quantity, Unit, Discount(if any), GST, Taxable Amount, Total, etc.

The bug of locked attribute of SheetText malfunctioning is resolved. The issue was very likely due to the g.dart file of sheet_text, it seemed to assign the default contructor value of locked to every sheetText on hot restart. Solved by making the attribute required. Other bug fixes and file cleanups