creds manager, edit time check and fetching bugs resolved.
the credentials are now managed in a class using riverpod and only signs in if the creds are missing or if they expired. Data is stored in gsheets and gdocs. Unliked last time this time the gdocs for spreadsheet is deleted each time a new one is pushed since the deleteConentRange doesnt seem to work for gDocs which led to errors which are now fixed. The modified dates of layouts and bills are compared and the user is prompted to resolve all the descrepancies. Implemented robust toMap and fromMap for each major model class